feat: introduce lazy-loading → item page
